Reduced % of bomber from 10-6 & bug fix
Now 6% chance of suicide bomber, down from 10% Also fixed a bug related
to FAR Revive system. FAR changes players to civilian side which meant
they had a chance of being equipped with a suicide vest. Added a check
to see if unit given the vest is player which then exits the script