Changed display using DynamicText instead of hint
Using Bohemia's dynamicText function, I removed the ugly hint system
(which was really just there for debugging and proof of concept in the
first place) and added the function to have text spawned in the lower
right side of the screen. This makes it noticeable but not obtrusive.
Text color, size, shadow, font and position are all extremely
configurable within the function itself. Also added more possibilities
for messages. Removed the distance format for Vehicle kills since the
distance would always be less than 3 anyway.