second pass on leaflet Drone
organized script for better readability (for anyone learning or looking
at it really), moved the IsDroneUp variable to a higher line for faster
check, didn't make sense to wait as long as I did before setting that.
shifted the "Reset Drone" function below the script and commented out
due to it needing more work. removed the Waypoint 2 statements because
we only used them for deleting the drone, which we do after a sleep
command anyway. Added what will be the "real" sleep timer and commented
it until we're ready for a Live test. added a call for the script to run
after its done.