added first version of drag and drop intel script
This intel script should be very useful for new designers as all they have to do is place these files in their root folder and place down  vanilla intel files down somewhere on the map.  The script will automatically find the intel files and apply the code to each. The number of possible files is endless 1 or 1,000,000+