version 1.0
A simple gate function on a laptop. There is some glitchy movement at the end of both Open and Close as I need to snap them to a designated position to prevent them from "migrating" during the move.

NOTE: If added to a road, some tweaking may be needed inorder to get a better visual on the gate bars themselves. This was built on a flat VR map.